+++ src/core/Authorization/Eve/AuthHandler.php
@@ -21,12 +21,13 @@
         public function __construct(
             protected $authorizationLogger,
             protected $authorizationConnection,
-            protected $authorizationVariables
+            protected $authorizationVariables,
+            protected $authorizationVersionVariables
         ) {
 
-            $this->esiHandler = new \Ridley\Objects\ESI\Handler($authorizationConnection);
+            $this->esiHandler = new \Ridley\Objects\ESI\Handler($this->authorizationConnection, $this->authorizationVersionVariables);
 
-            $this->cookieName = $authorizationVariables["Auth Cookie Name"];
+            $this->cookieName = $this->authorizationVariables["Auth Cookie Name"];
 
             $this->cleanupLogins();
             $this->cleanupSessions();
@@ -34,7 +35,7 @@
 
             if (isset($_GET["core_action"]) and $_GET["core_action"] == "login") {
 
-                $this->login("Default", $authorizationVariables["Default Scopes"]);
+                $this->login("Default", $this->authorizationVariables["Default Scopes"]);
 
             }
             if (isset($_GET["core_action"]) and $_GET["core_action"] == "callback") {
@@ -235,7 +236,7 @@
             $csrfBytes = random_bytes(16);
             $this->csrfToken = bin2hex($csrfBytes);
             $sessionExpiration = time() + $this->authorizationVariables["Session Time"];
-            setcookie($this->cookieName, $SessionID, ["expires" => $sessionExpiration, "path"=> "/", "samesite" => "Lax"]);
+            setcookie($this->cookieName, $SessionID, ["expires" => $sessionExpiration, "path"=> "/", "httponly" => true, "samesite" => "Lax"]);
 
             $convertedLoginStatus = (int)$this->isLoggedIn;
             $convertedAccessRoles = json_encode($this->accessRoles);